backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic-settings==2.2.1
scikit-learn==1.4.2
python-docx==1.1.0
pypdf==4.2.0
openai==1.30.1
